+++ matminer/featurizers/base.py
@@ -8,6 +8,7 @@ from six import string_types
 from multiprocessing import Pool, cpu_count
 
 from sklearn.base import TransformerMixin, BaseEstimator, is_classifier
+from matminer.utils.conversions import homogenize_multiindex
 
 
 class BaseFeaturizer(BaseEstimator, TransformerMixin):
@@ -155,7 +156,8 @@ class BaseFeaturizer(BaseEstimator, TransformerMixin):
                                                         **kwargs)
 
     def featurize_dataframe(self, df, col_id, ignore_errors=False,
-                            return_errors=False, inplace=True):
+                            return_errors=False, inplace=True,
+                            multiindex=False):
         """
         Compute features for all entries contained in input dataframe.
 
@@ -194,6 +196,15 @@ class BaseFeaturizer(BaseEstimator, TransformerMixin):
                                        return_errors=return_errors)
         if return_errors:
             labels.append(self.__class__.__name__ + " Exceptions")
+
+        if multiindex:
+            indices = ([self.__class__.__name__], labels)
+            labels = pd.MultiIndex.from_product(indices)
+            df = homogenize_multiindex(df, "Input Data")
+        elif isinstance(df.columns, pd.MultiIndex):
+            # If input df is multi, but multi not enabled...
+            raise ValueError("Please enable multiindexing to featurize an input"
+                             " dataframe containing a column multiindex.")
 
         # Create dataframe with the new features
         res = pd.DataFrame(features, index=df.index, columns=labels)
@@ -365,12 +376,24 @@ class MultipleFeaturizer(BaseFeaturizer):
     def feature_labels(self):
         return sum([f.feature_labels() for f in self.featurizers], [])
 
+    def fit_featurize_dataframe(self, df, col_id, *args, **kwargs):
+        for f in self.featurizers:
+            f.fit(df[col_id])
+        return self.featurize_dataframe(df, col_id, *args, **kwargs)
+
     def featurize_dataframe(self, df, col_id, ignore_errors=False,
-                            return_errors=False, inplace=True):
+                            return_errors=False, inplace=True,
+                            multiindex=False):
         """
         Featurize dataframe is overloaded in order to allow
         compatibility with Featurizers that overload featurize_dataframe
         """
+
+        if multiindex:
+            if not isinstance(df.columns, pd.MultiIndex):
+                col_id = ("Input Data", col_id)
+            df = homogenize_multiindex(df, "Input Data")
+
         # Detect if any featurizers override featurize_dataframe
         override = ["featurize_dataframe" in f.__class__.__dict__.keys()
                     for f in self.featurizers]
@@ -378,13 +401,16 @@ class MultipleFeaturizer(BaseFeaturizer):
             warnings.warn(
                 "One or more featurizers overrides featurize_dataframe, "
                 "featurization will be sequential and may diminish performance")
-            for f in self.featurizers:
-                df = f.featurize_dataframe(df, col_id, ignore_errors,
-                                           return_errors, inplace)
-                df[f.feature_labels()] = df[f.feature_labels()].applymap(np.squeeze)
-        else:
-            df = super(MultipleFeaturizer, self).featurize_dataframe(
-                df, col_id, ignore_errors, return_errors, inplace)
+
+        for f in self.featurizers:
+            df = f.featurize_dataframe(df, col_id, ignore_errors,
+                                       return_errors, inplace, multiindex)
+
+            if multiindex:
+                feature_labels = [(f.__class__.__name__, flabel) for flabel in f.feature_labels()]
+            else:
+                feature_labels = f.feature_labels()
+            df[feature_labels] = df[feature_labels].applymap(np.squeeze)
         return df
 
     def citations(self):