+++ matminer/featurizers/tests/test_site.py
@@ -8,10 +8,11 @@ from pymatgen.analysis.local_env import VoronoiNN, JMolNN
 
 from matminer.featurizers.site import AGNIFingerprints, \
     OPSiteFingerprint, CrystalNNFingerprint, \
-    CrystalSiteFingerprint, EwaldSiteEnergy, \
+    EwaldSiteEnergy, \
     VoronoiFingerprint, ChemEnvSiteFingerprint, \
     CoordinationNumber, ChemicalSRO, GaussianSymmFunc, \
     GeneralizedRadialDistributionFunction, AngularFourierSeries, LocalPropertyDifference
+from matminer.featurizers.deprecated import CrystalSiteFingerprint
 
 
 class FingerprintTests(PymatgenTest):
@@ -258,7 +259,7 @@ class FingerprintTests(PymatgenTest):
 
     def test_voronoifingerprint(self):
         df_sc= pd.DataFrame({'struct': [self.sc], 'site': [0]})
-        vorofp = VoronoiFingerprint(use_weights=True)
+        vorofp = VoronoiFingerprint(use_symm_weights=True)
         vorofps = vorofp.featurize_dataframe(df_sc, ['struct', 'site'])
         self.assertAlmostEqual(vorofps['Voro_index_3'][0], 0.0)
         self.assertAlmostEqual(vorofps['Voro_index_4'][0], 6.0)